pyproject.toml
[project]
name = "airesponder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agno>=2.2.10",
    "aiofiles>=25.1.0",
    "apache-airflow>=3.1.2",
    "chromadb>=1.3.4",
    "google-genai>=1.49.0",
    "groq>=0.33.0",
    "openai>=2.7.1",
    "pandas>=2.3.3",
    "prefect>=3.5.0",
    "sqlalchemy>=2.0.44",
]
